合约模块结构调整；新增README
